Create README.md
I have an extensive mainframe background which I walked away from when we ported our application to AIX. The rest of the application remained essentially Legacy. We're looking to port it to the cloud, transform it into an object oriented application, enable it for serious parallel processing and create a modern front-end to it. Looking fwd to a steep learning curve of a lot of new technologies.
Hobbies: been classically piano trained so I touch keyboards once in a while, do cycling, done skiing, like watching tennis, played some Ping-Pong, Chess, Backgamon and others